requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
requests==2.31.0
pandas==1.5.3  # 指定兼容版本
gunicorn==20.1.0
configparser==5.3.0
PyJWT==2.8.0
numpy==1.24.3
